[feat] 회원가입 액티비티 반응형 코드로 변환
