+++ app/src/main/java/org/sopt/sample/presentation/signup/view/SignUpActivity.kt
@@ -7,20 +7,14 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.widget.addTextChangedListener
 import com.google.android.material.snackbar.Snackbar
 import org.sopt.sample.R
-import org.sopt.sample.data.remote.api.AuthService
-import org.sopt.sample.data.remote.api.ServicePool
-import org.sopt.sample.data.remote.model.RequestSignUpDTO
-import org.sopt.sample.data.remote.model.ResponseSignUpDTO
 import org.sopt.sample.databinding.ActivitySignUpBinding
 import org.sopt.sample.presentation.login.view.LoginActivity
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import org.sopt.sample.presentation.signup.viewModel.SignUpViewModel
 
 
 class SignUpActivity : AppCompatActivity() {
     lateinit var binding: ActivitySignUpBinding
-    lateinit var loginService: AuthService
+    private val viewModel by lazy { SignUpViewModel() }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -29,6 +23,18 @@ class SignUpActivity : AppCompatActivity() {
 
         activateBtn()
         signUp()
+    }
+
+    private fun activateBtn() {
+        binding.idEt.addTextChangedListener {
+            setTextWatcher()
+        }
+        binding.pwET.addTextChangedListener {
+            setTextWatcher()
+        }
+        binding.etName.addTextChangedListener {
+            setTextWatcher()
+        }
     }
 
     private fun setTextWatcher() {
@@ -45,18 +51,6 @@ class SignUpActivity : AppCompatActivity() {
         }
     }
 
-    private fun activateBtn() {
-        binding.idEt.addTextChangedListener {
-            setTextWatcher()
-        }
-        binding.pwET.addTextChangedListener {
-            setTextWatcher()
-        }
-        binding.etName.addTextChangedListener {
-            setTextWatcher()
-        }
-    }
-
     private fun signUp() {
         binding.signupBtn.setOnClickListener {
             if (binding.idEt.text.length < 6) {
@@ -64,46 +58,23 @@ class SignUpActivity : AppCompatActivity() {
             } else if (binding.pwET.text.length !in 8..12) {
                 Snackbar.make(binding.root, "비밀번호는 8자 ~ 12자로 만들어주세요.", Snackbar.LENGTH_SHORT).show()
             } else {
-                loginService = ServicePool.authService
-                loginService.signUp(
-                    RequestSignUpDTO(
-                        binding.idEt.text.toString(),
-                        binding.pwET.text.toString(),
-                        binding.etName.text.toString()
-                    )
+                viewModel.signUp(
+                    binding.idEt.text.toString(),
+                    binding.pwET.toString(), binding.etName.toString()
                 )
-                    .enqueue(object : Callback<ResponseSignUpDTO> {
-                        override fun onResponse(
-                            call: Call<ResponseSignUpDTO>,
-                            response: Response<ResponseSignUpDTO>
-                        ) {
-                            if (response.isSuccessful) {
-                                signUpSuccess()
-                            } else {
-                                signUpBadResponse()
-                            }
-                        }
 
-                        override fun onFailure(call: Call<ResponseSignUpDTO>, t: Throwable) {
-                            signUpNoResponse()
-                        }
-                    })
+                viewModel.signUpResult.observe(this) {
+                    signUpSuccess()
+                }
+
+                viewModel.errorMessage.observe(this) {
+                    Toast.makeText(
+                        this@SignUpActivity,
+                        "회원가입 실패, 상태 코드 $it", Toast.LENGTH_SHORT
+                    ).show()
+                }
             }
         }
-    }
-
-    private fun signUpNoResponse() {
-        Toast.makeText(
-            this@SignUpActivity,
-            "네트워크 연결 미약", Toast.LENGTH_SHORT
-        ).show()
-    }
-
-    private fun signUpBadResponse() {
-        Toast.makeText(
-            this@SignUpActivity,
-            "회원가입 실패, 40X 응답값", Toast.LENGTH_SHORT
-        ).show()
     }
 
     private fun signUpSuccess() {
